requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
scipy>=1.7.0
Pillow>=8.0.0
tqdm>=4.60.0
